refactor : GameView 를 Input과 Output 기능에 맞게 분리
